refactor: centralize continuation ownership
